refactor: Moved writtable context into the validator module
